refactor(*): Split redis code to its own module
