+++ orca-core/orca-core.gradle
@@ -29,13 +29,11 @@ dependencies {
   compile spinnaker.dependency('cglib')
   compile spinnaker.dependency('jacksonGuava')
   compile spinnaker.dependency('rxJava')
-  compile spinnaker.dependency('jedis')
   compile spinnaker.dependency('korkSecurity')
   compile spinnaker.dependency('spectatorApi')
   compile spinnaker.dependency('kork')
   compile spinnaker.dependency('korkExceptions')
   compile spinnaker.dependency('korkArtifacts')
-  compile spinnaker.dependency('korkJedis')
   compile spinnaker.dependency("logstashEncoder")
   compile "com.fasterxml.jackson.datatype:jackson-datatype-jdk8:${spinnaker.version('jackson')}"
   compile "com.fasterxml.jackson.datatype:jackson-datatype-jsr310:${spinnaker.version('jackson')}"
@@ -46,5 +44,4 @@ dependencies {
 
   testCompile project(":orca-test")
   testCompile project(":orca-test-groovy")
-  testCompile spinnaker.dependency("korkJedisTest")
 }
