+++ orca-core/src/main/java/com/netflix/spinnaker/orca/notifications/NotificationClusterLock.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014 Netflix, Inc.
+ * Copyright 2018 Netflix, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,8 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package com.netflix.spinnaker.orca.notifications
 
-dependencies {
-  compile project(":orca-core")
-  compile spinnaker.dependency("spectatorApi")
+interface NotificationClusterLock {
+  fun tryAcquireLock(notificationType: String, lockTimeout: Long): Boolean
 }
