+++ orca-queue-tck/src/main/kotlin/com/netflix/spinnaker/orca/q/QueueIntegrationTest.kt
@@ -24,8 +24,6 @@ import com.netflix.spinnaker.assertj.assertSoftly
 import com.netflix.spinnaker.config.OrcaQueueConfiguration
 import com.netflix.spinnaker.config.QueueConfiguration
 import com.netflix.spinnaker.kork.eureka.RemoteStatusChangedEvent
-import com.netflix.spinnaker.kork.jedis.RedisClientDelegate
-import com.netflix.spinnaker.kork.jedis.RedisClientSelector
 import com.netflix.spinnaker.orca.CancellableStage
 import com.netflix.spinnaker.orca.ExecutionStatus.*
 import com.netflix.spinnaker.orca.TaskResult
@@ -45,11 +43,8 @@ import com.netflix.spinnaker.orca.pipeline.model.Stage
 import com.netflix.spinnaker.orca.pipeline.model.SyntheticStageOwner
 import com.netflix.spinnaker.orca.pipeline.model.SyntheticStageOwner.STAGE_BEFORE
 import com.netflix.spinnaker.orca.pipeline.persistence.ExecutionRepository
-import com.netflix.spinnaker.orca.pipeline.persistence.jedis.JedisConfiguration
-import com.netflix.spinnaker.orca.pipeline.persistence.jedis.RedisExecutionRepository
 import com.netflix.spinnaker.orca.pipeline.util.ContextParameterProcessor
 import com.netflix.spinnaker.orca.pipeline.util.StageNavigator
-import com.netflix.spinnaker.orca.test.redis.EmbeddedRedisConfiguration
 import com.netflix.spinnaker.q.DeadMessageCallback
 import com.netflix.spinnaker.q.Queue
 import com.netflix.spinnaker.q.memory.InMemoryQueue
@@ -72,12 +67,8 @@ import org.springframework.context.annotation.Configuration
 import org.springframework.context.annotation.Import
 import org.springframework.context.event.ApplicationEventMulticaster
 import org.springframework.context.event.SimpleApplicationEventMulticaster
-import org.springframework.scheduling.TaskScheduler
 import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor
-import org.springframework.scheduling.concurrent.ThreadPoolTaskScheduler
 import org.springframework.test.context.junit4.SpringRunner
-import redis.clients.jedis.Jedis
-import redis.clients.util.Pool
 import java.time.Clock
 import java.time.Duration
 import java.time.Instant.now
@@ -823,19 +814,14 @@ class QueueIntegrationTest {
 
 @Configuration
 @Import(
-  EmbeddedRedisConfiguration::class,
   PropertyPlaceholderAutoConfiguration::class,
   OrcaConfiguration::class,
   QueueConfiguration::class,
-  JedisConfiguration::class,
-  RedisExecutionRepository::class,
   StageNavigator::class,
   RestrictExecutionDuringTimeWindow::class,
   OrcaQueueConfiguration::class
 )
 class TestConfig {
-  @Bean
-  fun queueRedisPool(jedisPool: Pool<Jedis>) = jedisPool
 
   @Bean
   fun registry(): Registry = NoopRegistry()
@@ -926,10 +912,6 @@ class TestConfig {
       deadMessageHandlers = listOf(deadMessageHandler),
       publisher = publisher
     )
-
-  @Bean
-  fun redisClientSelector(redisClientDelegates: List<RedisClientDelegate>) =
-    RedisClientSelector(redisClientDelegates)
 
   @Bean
   fun applicationEventMulticaster(@Qualifier("applicationEventTaskExecutor") taskExecutor: ThreadPoolTaskExecutor): ApplicationEventMulticaster {
