+++ orca-redis/orca-redis.gradle
@@ -1,0 +1,12 @@
+apply from: "$rootDir/gradle/spock.gradle"
+
+dependencies {
+  compile project(":orca-front50")
+
+  compile spinnaker.dependency('rxJava')
+  compile spinnaker.dependency('jedis')
+  compile spinnaker.dependency('korkJedis')
+
+  testCompile project(":orca-core-tck")
+  testCompile project(":orca-test-redis")
+}
