+++ orca-redis/src/main/java/com/netflix/spinnaker/orca/config/RedisConnectionInfo.java
@@ -1,0 +1,78 @@
+/*
+ * Copyright 2018 Netflix, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.netflix.spinnaker.orca.config;
+
+import java.net.URI;
+import redis.clients.jedis.Protocol;
+import redis.clients.util.JedisURIHelper;
+
+public class RedisConnectionInfo {
+  public boolean hasPassword() {
+    return password.length() > 0;
+  }
+
+  public static RedisConnectionInfo parseConnectionUri(String connection) {
+
+    URI redisConnection = URI.create(connection);
+
+    String host = redisConnection.getHost();
+    int port = redisConnection.getPort() == -1 ? Protocol.DEFAULT_PORT : redisConnection.getPort();
+
+    int database = JedisURIHelper.getDBIndex(redisConnection);
+
+    String password = JedisURIHelper.getPassword(redisConnection);
+
+    return new RedisConnectionInfo();
+  }
+
+  public String getHost() {
+    return host;
+  }
+
+  public void setHost(String host) {
+    this.host = host;
+  }
+
+  public int getPort() {
+    return port;
+  }
+
+  public void setPort(int port) {
+    this.port = port;
+  }
+
+  public int getDatabase() {
+    return database;
+  }
+
+  public void setDatabase(int database) {
+    this.database = database;
+  }
+
+  public String getPassword() {
+    return password;
+  }
+
+  public void setPassword(String password) {
+    this.password = password;
+  }
+
+  private String host;
+  private int port;
+  private int database;
+  private String password;
+}
+
