+++ orca-redis/src/main/java/com/netflix/spinnaker/orca/notifications/RedisNotificationClusterLock.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2018 Netflix, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.netflix.spinnaker.orca.notifications;
+
+import com.netflix.spinnaker.kork.jedis.RedisClientDelegate;
+
+import javax.annotation.Nonnull;
+
+public class RedisNotificationClusterLock implements NotificationClusterLock {
+
+  private final RedisClientDelegate redisClientDelegate;
+
+  public RedisNotificationClusterLock(RedisClientDelegate redisClientDelegate) {
+    this.redisClientDelegate = redisClientDelegate;
+  }
+
+  @Override
+  public boolean tryAcquireLock(@Nonnull String notificationType, long lockTimeout) {
+    String key = "lock:" + notificationType;
+    return redisClientDelegate.withCommandsClient(client -> {
+      return "OK".equals(client.set(key, "\uD83D\uDD12", "NX", "EX", lockTimeout));
+    });
+  }
+}
