+++ orca-test-redis/src/main/java/com/netflix/spinnaker/orca/test/redis/EmbeddedRedisConfiguration.java
